Move some methods to packages.go
