+++ internal/initrec/initrec.go
@@ -1,5 +1,6 @@
 // Copyright 2023 Jetpack Technologies Inc and contributors. All rights reserved.
 // Use of this source code is governed by the license in the LICENSE file.
+
 package initrec
 
 import (
@@ -18,7 +19,6 @@ import (
 )
 
 func getRecommenders(srcDir string) []recommenders.Recommender {
-
 	return []recommenders.Recommender{
 		&dotnet.Recommender{SrcDir: srcDir},
 		&golang.Recommender{SrcDir: srcDir},
