+++ internal/initrec/recommenders/interface.go
@@ -1,5 +1,6 @@
 // Copyright 2023 Jetpack Technologies Inc and contributors. All rights reserved.
 // Use of this source code is governed by the license in the LICENSE file.
+
 package recommenders
 
 type Recommender interface {
